Allow Anony to be configured to ignore certain columns
In our application at least, the :id, :created_at and :updated_at
columns are all usefully used elsewhere and we don't want to apply
anonymisation rules. Similarly, the links to other models (e.g.
object.organisation_id) should be preserved because we want to be able
to traverse those relations later even if the original object is
anonymised.

This commit adds support for this feature, makes it configurable and
disabled by default, and prevents "ignoring" fields which would already
be ignored (however, it is deliberately still possible to define an
anonymisation strategy anyway for one of these columns if the
application requires it).