Give Anony support for multi-fields
A nice addition to the interface is supporting multiple fields with a
single block, e.g.:

  anonymise do
    with_strategy(:a, :b, :c) { |v| v.reverse }
  end

Because it's not possible to have a splat argument, followed by a
default then a block, this commit alters `#with_strategy` to accept the
strategy object as the first argument followed by one or more fields.
Naturally that interface is more complex to support internally so there
are some unit tests on that method to ensure it has at least one
strategy and at least one field given.

This commit also changes emails to end with @example.com

If we use @gocardless.com and then end up trying to send this email (a
not-inconceivable possibility given the myriad ways we try to contact
users), it will hit our own mail servers and then be discarded or
logged.

Instead, pointing the email at @example.com will blackhole the email.
Email servers recognise this is a dummy domain and simply won't proceed,
saving the overhead of actually trying to deliver the message.