Use anonymous classes in Anony spec
We've seen a few flaky specs [1] from this file and this is potentially
happening because of the way that RSpec works and the file is parsed.

Essentially, defining a class in an RSpec test doesn't coerce it to the
same namespace (so here, StubModel would be defined in the top-level
scope) and it also doesn't obey the usual rules of precedence.

In rare cases, depending on the way the file was loaded and the order
the specs were executed, the InvalidModel class is actually defined in
several places and so behaviour could be overridden during the spec
execution.

Moving to anonymous classes (Class.new { .. }) allows us to safely
attach behaviour whilst preventing collision between specs. The syntax
is a little ugly but it gives us the isolation guarantees we were
otherwise expecting to get.

[1]: https://gocardless.myjetbrains.com/youtrack/issue/TX-5976