Prevent anonymisation spec from blatting global configuration
We were seeing some flaky tests that happen after this spec runs,
because this spec is changing the global state that other specs will be
relying on.

Instead, we can just use this around block to ensure that things are put
back correctly after the spec. If we ever extract this as a library, we
can remove this fixup.