GDPR anon
Ref [MIXP-270](https://gocardless.atlassian.net/browse/MIXP-270)

As per [our GDPR
policy](https://docs.google.com/spreadsheets/d/1Ol-wbz6ZX3aJN0dGi12sYhMgsHcfjkngRAIcAkUsfcc/edit#gid=0&fvid=1142575516),
we need to anonymise any personal fields on models owned by MIXP.

The fields that we're responsible for anonymising are the `phone_number`
& `metadata` fields on the `AuthenticationFactor` model.

Note, for the phone number field, we've decided to anonymise it with a real phone number
that we control. The rationale behind this is that we want to be sure,
beyond a doubt, that any messages sent to the anonymised number are
not being leaked to a real number by accident.