Allow accessing model methods when using block anonymisers
instance_exec allows us to execute a block, with arguments, in the local
context, allowing the block to access local behaviours (for example,
your anonymisation could be conditional, or rely on another property of
the model).

We don't support this for Modules/etc at this time because it would
require converting the module into a proc first (via
Module.method(:call)), which has performance and behavioural
implications.