Only apply anonymisation strategies to columns that are defined
Currently, anonymisation strategies have to be kept in lockstep with the
database structure.  This is a problem when code deploys and database
migrations happen at different times.

This commit ensures that anonymisation strategies only run for columns
that are defined in the databse.  This means that anonymisation
strategies can be defined and deployed before database migrations run,
and they will take effect once the new column is in place.