Prevent multiple strategies on a field
Its possible currently to define an anony config with multiple
strategies for a field, however the last strategy is the one used. This
can be problematic if this is the NoOp strategy.

For example

```ruby
anonmise do
  nilable :foo
  ignore :foo
end
```

Would result in a No-op, this PR introduces an error when this mapping
is encountered.