Introduce before/after and around hooks
In some cases it's useful to perform some extra actions when we're about
to anonimise an object, e.g. in case of models based on event sourcing
it may be desirable to emit an event indicating that the model has been
anonymised alongside the anonimisation itself. Or perhaps one may want
to (temporarily) allow write actions on some field that are about to be
anonymised on models that are read-only by default. Or maybe even just
add metrics around anonymisations.

This commit introduces before/after and around hooks to facilitate all
of these use cases (and perhaps some more).