Move all strategies into a single file
In preparation for an upcoming refactor, the Strategies namespace is
going to be reused for a higher-level concept. This moves strategies
relating to fields into a special module, FieldLevelStrategies.

This commit also moves all of the strategies and their specs into a
single file. This is because the definition of strategies should be
effectively atomic with the loading of the dynamic registration code.
This means that consumers which require this module will automatically
get all of the necessary strategies configured for them without having
to load separate files.