Use ActiveRecord::Persistence#current_time_from_proper_timezone
This method is used internally by the #touch interface, and uses some
intelligent delegation to figure out which TimeZone to set. This also
means that you shouldn't need to configure `Time.zone` before
anonymising any model (only if you want to use that particular
strategy).

Note that this changes the dependencies of the gem - before, you could
reimplement a few of the core ActiveRecord methods, now this increases
the coupling with that library.