Update dependencies to drop Ruby 3.0 and fix dependabot
Dependabot has not been running successfully in this repository due to
the way we were specifying the Rails versions in the Gemfile. We were
using the Gemfile to set the Rails versions from an environment variable
to allow us to test against specific versions, but Dependabot was
checking the Gemfile and, not having those env vars, was failing to
determine an appropriate version to use and erroring.

The Rails version specifications have now been moved fully to the
gemspec and are configured so that if no env var is present to set a
specific Rails version, it will use the version range that was
previously specified in the gemspec.

Also drop support for EOL Ruby 3.0 and add Rails 7.2 to the test matrix
to ensure compatibility