Include reference to anonymised record in `Anony::Result`
When using selectors to anonymise records, an array of `Result` objects
is returned and these results now contain a reference to the record that
was anonymised.

This means that it is significantly easier to match the changes from the
`fields` on each result with the record that those changes came from in
case the changes don't include any unique identifiers like the primary
key field.