Add `ignore` strategy that skips anonymisation of a model
It is not uncommon for models to contain no personal data that needs
anonymising and this is often worked around by listing every field in
the model under an `ignore` field strategy as part of an `overwrite`
strategy.

To simplify config for these situations, add an `ignore` strategy that
causes the model to be skipped when anonymising.

Also add a helper method `requires_anonymisation?` to make it easy to
tell whether or not a model needs anonymising, which can save time and
DB queries when finding records to be anonymised